+++ tests/test_api.py
@@ -3,7 +3,7 @@ import re
 import shutil
 import time
 
-from httprunner import exceptions, loader, parser
+from httprunner import exceptions, loader, parser, report
 from httprunner.api import HttpRunner
 from tests.api_server import HTTPBIN_SERVER
 from tests.base import ApiServerUnittest
@@ -34,7 +34,8 @@ class TestHttpRunner(ApiServerUnittest):
                     'request': {
                         'url': 'http://127.0.0.1:5000/api/get-token',
                         'method': 'POST',
-                        'headers': {'Content-Type': 'application/json', 'app_version': '2.8.6', 'device_sn': 'FwgRiO7CNA50DSU', 'os_platform': 'ios', 'user_agent': 'iOS/10.3'},
+                        'headers': {'Content-Type': 'application/json', 'app_version': '2.8.6',
+                                    'device_sn': 'FwgRiO7CNA50DSU', 'os_platform': 'ios', 'user_agent': 'iOS/10.3'},
                         'json': {'sign': '9c0c7e51c91ae963c833a4ccbab8d683c4a90c98'}
                     },
                     'extract': [
@@ -51,7 +52,8 @@ class TestHttpRunner(ApiServerUnittest):
                     'request': {
                         'url': 'http://127.0.0.1:5000/api/users/1000',
                         'method': 'POST',
-                        'headers': {'Content-Type': 'application/json', 'device_sn': 'FwgRiO7CNA50DSU','token': '$token'},
+                        'headers': {'Content-Type': 'application/json',
+                                    'device_sn': 'FwgRiO7CNA50DSU','token': '$token'},
                         'json': {'name': 'user1', 'password': '123456'}
                     },
                     'validate': [
@@ -75,9 +77,9 @@ class TestHttpRunner(ApiServerUnittest):
         return self.api_client.get(url, headers=headers)
 
     def test_text_run_times(self):
-        self.runner.run(self.testcase_cli_path)
-        self.assertEqual(self.runner.summary["stat"]["testcases"]["total"], 1)
-        self.assertEqual(self.runner.summary["stat"]["teststeps"]["total"], 10)
+        summary = self.runner.run(self.testcase_cli_path)
+        self.assertEqual(summary["stat"]["testcases"]["total"], 1)
+        self.assertEqual(summary["stat"]["teststeps"]["total"], 10)
 
     def test_text_run_times_invalid(self):
         testcases = [
@@ -113,8 +115,8 @@ class TestHttpRunner(ApiServerUnittest):
             self.runner.run_tests(tests_mapping)
 
     def test_text_skip(self):
-        self.runner.run(self.testcase_cli_path)
-        self.assertEqual(self.runner.summary["stat"]["teststeps"]["skipped"], 4)
+        summary = self.runner.run(self.testcase_cli_path)
+        self.assertEqual(summary["stat"]["teststeps"]["skipped"], 4)
 
     def test_save_variables_output(self):
         testcases = [
@@ -179,27 +181,29 @@ class TestHttpRunner(ApiServerUnittest):
         self.assertEqual(token1, token2)
 
     def test_html_report(self):
-        report_save_dir = os.path.join(os.getcwd(), 'reports', "demo")
-        runner = HttpRunner(failfast=True, report_dir=report_save_dir)
-        runner.run(self.testcase_cli_path)
-        summary = runner.summary
+        runner = HttpRunner(failfast=True)
+        summary = runner.run(self.testcase_cli_path)
         self.assertEqual(summary["stat"]["testcases"]["total"], 1)
         self.assertEqual(summary["stat"]["teststeps"]["total"], 10)
         self.assertEqual(summary["stat"]["teststeps"]["skipped"], 4)
+
+        report_save_dir = os.path.join(os.getcwd(), 'reports', "demo")
+        report.gen_html_report(summary, report_dir=report_save_dir)
         self.assertGreater(len(os.listdir(report_save_dir)), 0)
         shutil.rmtree(report_save_dir)
 
     def test_html_report_with_fixed_report_file(self):
-        report_save_dir = os.path.join(os.getcwd(), 'reports', "demo")
-        report_file = 'test.html'
-        runner = HttpRunner(failfast=True, report_dir=report_save_dir, report_file=report_file)
-        runner.run(self.testcase_cli_path)
-        summary = runner.summary
+        runner = HttpRunner(failfast=True)
+        summary = runner.run(self.testcase_cli_path)
         self.assertEqual(summary["stat"]["testcases"]["total"], 1)
         self.assertEqual(summary["stat"]["teststeps"]["total"], 10)
         self.assertEqual(summary["stat"]["teststeps"]["skipped"], 4)
+
+        report_file = os.path.join(os.getcwd(), 'reports', "demo", "test.html")
+        report.gen_html_report(summary, report_file=report_file)
+        report_save_dir = os.path.dirname(report_file)
         self.assertEqual(len(os.listdir(report_save_dir)), 1)
-        self.assertTrue(os.path.isfile(os.path.join(report_save_dir, report_file)))
+        self.assertTrue(os.path.isfile(report_file))
         shutil.rmtree(report_save_dir)
 
     def test_log_file(self):
@@ -210,8 +214,7 @@ class TestHttpRunner(ApiServerUnittest):
         os.remove(log_file_path)
 
     def test_run_testcases(self):
-        self.runner.run_tests(self.tests_mapping)
-        summary = self.runner.summary
+        summary = self.runner.run_tests(self.tests_mapping)
         self.assertTrue(summary["success"])
         self.assertEqual(summary["stat"]["testcases"]["total"], 1)
         self.assertEqual(summary["stat"]["teststeps"]["total"], 2)
@@ -219,8 +222,7 @@ class TestHttpRunner(ApiServerUnittest):
         self.assertIn("records", summary["details"][0])
 
     def test_run_yaml_upload(self):
-        self.runner.run("tests/httpbin/upload.yml")
-        summary = self.runner.summary
+        summary = self.runner.run("tests/httpbin/upload.yml")
         self.assertTrue(summary["success"])
         self.assertEqual(summary["stat"]["testcases"]["total"], 1)
         self.assertEqual(summary["stat"]["teststeps"]["total"], 1)
@@ -256,8 +258,7 @@ class TestHttpRunner(ApiServerUnittest):
         tests_mapping = {
             "testcases": testcases
         }
-        self.runner.run_tests(tests_mapping)
-        summary = self.runner.summary
+        summary = self.runner.run_tests(tests_mapping)
         self.assertTrue(summary["success"])
         self.assertEqual(summary["stat"]["testcases"]["total"], 1)
         self.assertEqual(summary["stat"]["teststeps"]["total"], 1)
@@ -267,23 +268,23 @@ class TestHttpRunner(ApiServerUnittest):
         )
 
     def test_html_report_repsonse_image(self):
+        runner = HttpRunner(failfast=True)
+        summary = runner.run("tests/httpbin/load_image.yml")
+
         report_save_dir = os.path.join(os.getcwd(), 'reports', "demo")
-        runner = HttpRunner(failfast=True, report_dir=report_save_dir)
-        report = runner.run("tests/httpbin/load_image.yml")
-        self.assertTrue(os.path.isfile(report))
+        report_path = report.gen_html_report(summary, report_dir=report_save_dir)
+        self.assertTrue(os.path.isfile(report_path))
         shutil.rmtree(report_save_dir)
 
     def test_testcase_layer_with_api(self):
-        self.runner.run("tests/testcases/setup.yml")
-        summary = self.runner.summary
+        summary = self.runner.run("tests/testcases/setup.yml")
         self.assertTrue(summary["success"])
         self.assertEqual(summary["details"][0]["records"][0]["name"], "get token (setup)")
         self.assertEqual(summary["stat"]["testcases"]["total"], 1)
         self.assertEqual(summary["stat"]["teststeps"]["total"], 2)
 
     def test_testcase_layer_with_testcase(self):
-        self.runner.run("tests/testsuites/create_users.yml")
-        summary = self.runner.summary
+        summary = self.runner.run("tests/testsuites/create_users.yml")
         self.assertTrue(summary["success"])
         self.assertEqual(summary["stat"]["testcases"]["total"], 2)
         self.assertEqual(summary["stat"]["teststeps"]["total"], 4)
@@ -292,9 +293,8 @@ class TestHttpRunner(ApiServerUnittest):
         testcase_file_path = os.path.join(
             os.getcwd(), 'tests/httpbin/hooks.yml')
         start_time = time.time()
-        self.runner.run(testcase_file_path)
+        summary = self.runner.run(testcase_file_path)
         end_time = time.time()
-        summary = self.runner.summary
         self.assertTrue(summary["success"])
         self.assertLess(end_time - start_time, 60)
 
@@ -332,8 +332,7 @@ class TestHttpRunner(ApiServerUnittest):
             "project_mapping": loader.project_mapping,
             "testcases": testcases
         }
-        self.runner.run_tests(tests_mapping)
-        summary = self.runner.summary
+        summary = self.runner.run_tests(tests_mapping)
         self.assertTrue(summary["success"])
 
     def test_run_httprunner_with_teardown_hooks_not_exist_attribute(self):
@@ -365,8 +364,7 @@ class TestHttpRunner(ApiServerUnittest):
             "project_mapping": loader.project_mapping,
             "testcases": testcases
         }
-        self.runner.run_tests(tests_mapping)
-        summary = self.runner.summary
+        summary = self.runner.run_tests(tests_mapping)
         self.assertFalse(summary["success"])
         self.assertEqual(summary["stat"]["teststeps"]["errors"], 1)
 
@@ -396,15 +394,13 @@ class TestHttpRunner(ApiServerUnittest):
             "project_mapping": loader.project_mapping,
             "testcases": testcases
         }
-        self.runner.run_tests(tests_mapping)
-        summary = self.runner.summary
+        summary = self.runner.run_tests(tests_mapping)
         self.assertFalse(summary["success"])
         self.assertEqual(summary["stat"]["teststeps"]["errors"], 1)
 
     def test_run_api(self):
         path = "tests/httpbin/api/get_headers.yml"
-        self.runner.run(path)
-        summary = self.runner.summary
+        summary = self.runner.run(path)
         self.assertTrue(summary["success"])
         self.assertEqual(summary["stat"]["testcases"]["total"], 1)
         self.assertEqual(summary["stat"]["teststeps"]["total"], 1)
@@ -412,8 +408,7 @@ class TestHttpRunner(ApiServerUnittest):
 
     def test_request_302_logs(self):
         path = "tests/httpbin/api/302_redirect.yml"
-        self.runner.run(path)
-        summary = self.runner.summary
+        summary = self.runner.run(path)
         self.assertTrue(summary["success"])
         self.assertEqual(summary["stat"]["testcases"]["total"], 1)
         self.assertEqual(summary["stat"]["teststeps"]["total"], 1)
@@ -426,8 +421,7 @@ class TestHttpRunner(ApiServerUnittest):
 
     def test_request_with_params(self):
         path = "tests/httpbin/api/302_redirect.yml"
-        self.runner.run(path)
-        summary = self.runner.summary
+        summary = self.runner.run(path)
         self.assertTrue(summary["success"])
         self.assertEqual(summary["stat"]["testcases"]["total"], 1)
         self.assertEqual(summary["stat"]["teststeps"]["total"], 1)
@@ -442,8 +436,7 @@ class TestHttpRunner(ApiServerUnittest):
 
     def test_run_api_folder(self):
         api_folder = "tests/httpbin/api/"
-        self.runner.run(api_folder)
-        summary = self.runner.summary
+        summary = self.runner.run(api_folder)
         self.assertTrue(summary["success"])
         self.assertEqual(summary["stat"]["testcases"]["total"], 2)
         self.assertEqual(summary["stat"]["teststeps"]["total"], 2)
@@ -454,8 +447,7 @@ class TestHttpRunner(ApiServerUnittest):
 
     def test_run_testcase_hardcode(self):
         for testcase_file_path in self.testcase_file_path_list:
-            self.runner.run(testcase_file_path)
-            summary = self.runner.summary
+            summary = self.runner.run(testcase_file_path)
             self.assertTrue(summary["success"])
             self.assertEqual(summary["stat"]["testcases"]["total"], 1)
             self.assertEqual(summary["stat"]["teststeps"]["total"], 3)
@@ -465,30 +457,26 @@ class TestHttpRunner(ApiServerUnittest):
     def test_run_testcase_template_variables(self):
         testcase_file_path = os.path.join(
             os.getcwd(), 'tests/data/demo_testcase_variables.yml')
-        self.runner.run(testcase_file_path)
-        summary = self.runner.summary
+        summary = self.runner.run(testcase_file_path)
         self.assertTrue(summary["success"])
 
     def test_run_testcase_template_import_functions(self):
         testcase_file_path = os.path.join(
             os.getcwd(), 'tests/data/demo_testcase_functions.yml')
-        self.runner.run(testcase_file_path)
-        summary = self.runner.summary
+        summary = self.runner.run(testcase_file_path)
         self.assertTrue(summary["success"])
 
     def test_run_testcase_layered(self):
         testcase_file_path = os.path.join(
             os.getcwd(), 'tests/data/demo_testcase_layer.yml')
-        self.runner.run(testcase_file_path)
-        summary = self.runner.summary
+        summary = self.runner.run(testcase_file_path)
         self.assertTrue(summary["success"])
         self.assertEqual(len(summary["details"]), 1)
 
     def test_run_testcase_output(self):
         testcase_file_path = os.path.join(
             os.getcwd(), 'tests/data/demo_testcase_layer.yml')
-        self.runner.run(testcase_file_path)
-        summary = self.runner.summary
+        summary = self.runner.run(testcase_file_path)
         self.assertTrue(summary["success"])
         self.assertIn("token", summary["details"][0]["in_out"]["out"])
         # TODO: add
@@ -500,8 +488,7 @@ class TestHttpRunner(ApiServerUnittest):
         variables_mapping = {
             "app_version": '2.9.7'
         }
-        self.runner.run(testcase_file_path, mapping=variables_mapping)
-        summary = self.runner.summary
+        summary = self.runner.run(testcase_file_path, mapping=variables_mapping)
         self.assertTrue(summary["success"])
         self.assertIn("token", summary["details"][0]["in_out"]["out"])
         # TODO: add
@@ -510,8 +497,7 @@ class TestHttpRunner(ApiServerUnittest):
     def test_run_testcase_with_parameters(self):
         testcase_file_path = os.path.join(
             os.getcwd(), 'tests/testsuites/create_users_with_parameters.yml')
-        self.runner.run(testcase_file_path)
-        summary = self.runner.summary
+        summary = self.runner.run(testcase_file_path)
         self.assertTrue(summary["success"])
         self.assertEqual(len(summary["details"]), 3 * 2)
 
@@ -549,8 +535,8 @@ class TestHttpRunner(ApiServerUnittest):
     #     # TODO: fix compatibility with Python 2.7
     #     testcase_file_path = os.path.join(
     #         os.getcwd(), 'tests/httpbin/basic.yml')
-    #     self.runner.run(testcase_file_path)
-    #     self.assertTrue(self.runner.summary["success"])
+    #     summary = self.runner.run(testcase_file_path)
+    #     self.assertTrue(summary["success"])
 
     def test_html_report_xss(self):
         testcases = [
@@ -582,7 +568,8 @@ class TestHttpRunner(ApiServerUnittest):
         tests_mapping = {
             "testcases": testcases
         }
-        report_path = self.runner.run(tests_mapping)
+        summary = self.runner.run(tests_mapping)
+        report_path = report.gen_html_report(summary)
         with open(report_path) as f:
             content = f.read()
             m = re.findall(